add a multithreaded inflate implementation
Move the single-threaded version inherited from p7zip to DeflateDecoderST,
create a multithreaded one in DeflateDecoderMT, and make DeflateDecoder
be a wrapper around both, choosing its implementation depending on
it "multithreaded" or "threadCount" properties.

The multithreaded inflate works by having one thread parse the bit stream
to produce a set of literals and (distance, length) pairs, and having a
second thread take that stream and apply it to the output buffer. It
seems to be worth ~20% in wall time on a single-CPU HT Pentium4... but
we're still much slower than p7zip or gzip. :-(